% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyin.R
\name{test_pyin}
\alias{test_pyin}
\title{Test pyin is working}
\usage{
test_pyin(if_bad_result_return_single_na = TRUE)
}
\arguments{
\item{if_bad_result_return_single_na}{(logical scalar) Whether or not to
return a single \code{NA} instead of a tibble if pYIN execution has a bad
result.}
}
\value{
If the pYIN setup was successful on your system, the function
returns a 10x5 tibble with the transcribed note events of the test file.
}
\description{
Test pyin is working
}
