% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyin.R
\name{pyin}
\alias{pyin}
\title{Compute pYIN on an audio track}
\usage{
pyin(
  file_name,
  transform_file = NULL,
  normalise = FALSE,
  hidePrint = TRUE,
  type = c("notes", "pitch_track", "both"),
  if_bad_result_return_single_na = TRUE
)
}
\arguments{
\item{if_bad_result_return_single_na}{}
}
\value{

}
\description{
Compute pYIN on an audio track
}
