% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyin.R
\name{test_pyin}
\alias{test_pyin}
\title{Test pyin is working}
\usage{
test_pyin(if_bad_result_return_single_na = TRUE)
}
\arguments{
\item{if_bad_result_return_single_na}{}
}
\value{

}
\description{
Test pyin is working
}
